+++ contracts/test/CourtMock.sol
@@ -9,7 +9,7 @@ contract CourtMock is Court {
     bool internal treeSearchHijacked = false;
 
     constructor(
-        uint64 _termDuration,
+        uint32 _termDuration,
         ERC20 _jurorToken,
         ERC20 _feeToken,
         uint256 _jurorFee,
@@ -20,7 +20,7 @@ contract CourtMock is Court {
         address _governor,
         uint64 _firstTermStartTime,
         uint256 _jurorMinStake,
-        uint64[3] _roundStateDurations,
+        uint32[3] _roundStateDurations,
         uint16 _penaltyPct
     ) Court(
         _termDuration,
